requirements.txt
requests
apscheduler
Flask-API.yandex
flask
# faster_than_requests==0.9.0
prettytable
gmail
